053 - Longest increasing sub-sequence
Coaching another developer through solving this challenge.  Started with loop
recur to show how complexe the code would be and didnt quite get a working
solution.

Switched to a higher abstraction with partition and with the help of filter and
sort-by was able to create a much nicer solution, which actually did work.