022 - Count a sequence - some lateral thinking going on
As we are restricted from using the `count` function, then some lateral thinking
was applied to come up with some interesting answers.

Included solutions using loop/recur, recursive function, reduce/or, and
map/constantly

The reduce/or and map/constantly did not come easily and took a fair bit of
lateral thinking.